feat: avoid recreate fileOptions and clickOptionsHandler for each file node
