+++ src/components/file-item/file-item.tsx
@@ -1,24 +1,12 @@
 import {
+    ConnectDropdownMenuItem,
     FileItem as ConnectFileItem,
-    FileItemProps,
-    Icon,
     TreeItem,
-    defaultDropdownMenuOptions,
 } from '@powerhousedao/design-system';
 import { useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useDocumentDriveById } from 'src/hooks/useDocumentDriveById';
 import { useDrivesContainer } from 'src/hooks/useDrivesContainer';
-import { useGetDocumentById } from 'src/hooks/useGetDocumentById';
+import { SetIsWriteMode } from 'src/hooks/useFileOptions';
 import { useGetReadableItemPath } from 'src/hooks/useGetReadableItemPath';
-import { useOpenSwitchboardLink } from 'src/hooks/useOpenSwitchboardLink';
-import { useModal } from '../modal';
-
-const allowedItemOptions = ['delete', 'rename', 'duplicate'];
-
-const defaultItemOptions = defaultDropdownMenuOptions.filter(option =>
-    allowedItemOptions.includes(option.id),
-);
 
 interface IProps {
     file: TreeItem;
@@ -26,67 +14,25 @@ interface IProps {
     onFileSelected: (drive: string, id: string) => void;
     onFileDeleted: (drive: string, id: string) => void;
     isAllowedToCreateDocuments: boolean;
+    itemOptions?: ConnectDropdownMenuItem[];
+    onFileOptionsClick?: (
+        optionId: string,
+        fileNode: TreeItem,
+        setIsWriteMode: SetIsWriteMode,
+    ) => Promise<void>;
 }
 
 export const FileItem: React.FC<IProps> = ({
     file,
     decodedDriveID,
     onFileSelected,
+    itemOptions = [],
     isAllowedToCreateDocuments,
+    onFileOptionsClick = () => {},
 }) => {
-    const { t } = useTranslation();
     const [isWriteMode, setIsWriteMode] = useState(false);
     const getReadableItemPath = useGetReadableItemPath();
-    const getDocumentById = useGetDocumentById();
-    const { updateNodeName, onSubmitInput } = useDrivesContainer();
-    const { showModal } = useModal();
-
-    const openSwitchboardLink = useOpenSwitchboardLink(decodedDriveID);
-    const { isRemoteDrive } = useDocumentDriveById(decodedDriveID);
-
-    // TODO: move this to folder-view
-    const onFileOptionsClick = async (optionId: string, fileNode: TreeItem) => {
-        if (optionId === 'delete') {
-            showModal('deleteItem', {
-                driveId: decodedDriveID,
-                itemId: fileNode.id,
-                itemName: file.label,
-                type: 'file',
-            });
-        }
-        if (optionId === 'duplicate') {
-            await onSubmitInput({
-                ...file,
-                action: 'UPDATE_AND_COPY',
-            });
-        }
-
-        if (optionId === 'rename') {
-            setIsWriteMode(true);
-        }
-
-        if (optionId === 'switchboard-link') {
-            const document = getDocumentById(decodedDriveID, fileNode.id);
-
-            await openSwitchboardLink(document);
-        }
-    };
-
-    // TODO: move this to folder-view
-    const itemOptions: FileItemProps['itemOptions'] = isRemoteDrive
-        ? [
-              {
-                  id: 'switchboard-link',
-                  label: t('files.options.switchboardLink'),
-                  icon: (
-                      <div className="flex w-6 justify-center">
-                          <Icon name="drive" size={16} />
-                      </div>
-                  ),
-              },
-              ...defaultItemOptions,
-          ]
-        : defaultItemOptions;
+    const { updateNodeName } = useDrivesContainer();
 
     const cancelInputHandler = () => setIsWriteMode(false);
     const submitInputHandler = (value: string) => {
@@ -105,7 +51,9 @@ export const FileItem: React.FC<IProps> = ({
             onCancelInput={cancelInputHandler}
             onSubmitInput={submitInputHandler}
             mode={isWriteMode ? 'write' : 'read'}
-            onOptionsClick={optionId => onFileOptionsClick(optionId, file)}
+            onOptionsClick={optionId =>
+                onFileOptionsClick(optionId, file, setIsWriteMode)
+            }
             item={file}
             onClick={() =>
                 !isWriteMode && onFileSelected(decodedDriveID, file.id)
